unify audio and video processing to single queue
